+++ index.js
@@ -28,8 +28,6 @@ function withEach(target, cb) {
   })(target);
 }
 
-// target: 'prohibited.key': 'value',
-// allowDots: true
 function has(target, allowDots) {
   const regex = getTestRegex(allowDots);
 
@@ -46,13 +44,12 @@ function has(target, allowDots) {
   return hasProhibited;
 }
 
-function sanitize(target, options) {
-  options = options || {};
-
-  // Regex is not passed from the middleware
+function _sanitize(target, options) {
   const regex = getTestRegex(options.allowDots);
 
+  let isSanitized = false;
   let replaceWith = null;
+  const dryRun = Boolean(options.dryRun);
   if (!regex.test(options.replaceWith) && options.replaceWith !== '.') {
     replaceWith = options.replaceWith;
   }
@@ -61,13 +58,25 @@ function sanitize(target, options) {
     let shouldRecurse = true;
 
     if (regex.test(key)) {
+      isSanitized = true;
+      // if dryRun is enabled, do not modify the target
+      if (dryRun) {
+        return {
+          shouldRecurse: shouldRecurse,
+          key: key,
+        };
+      }
       delete obj[key];
       if (replaceWith) {
         key = key.replace(REPLACE_REGEX, replaceWith);
         // Avoid to set __proto__ and constructor.prototype
         // https://portswigger.net/daily-swig/prototype-pollution-the-dangerous-and-underrated-vulnerability-impacting-javascript-applications
         // https://snyk.io/vuln/SNYK-JS-LODASH-73638
-        if (key !== '__proto__' && key !== 'constructor' && key !== 'prototype') {
+        if (
+          key !== '__proto__' &&
+          key !== 'constructor' &&
+          key !== 'prototype'
+        ) {
           obj[key] = val;
         }
       } else {
@@ -81,14 +90,33 @@ function sanitize(target, options) {
     };
   });
 
-  return target;
+  return {
+    isSanitized,
+    target,
+  };
 }
 
-function middleware(options) {
+function sanitize(target, options) {
+  return _sanitize(target, options).target;
+}
+
+/**
+ * @param {{replaceWith?: string, onSanitize?: function, dryRun?: boolean}} options
+ * @returns {function}
+ */
+function middleware(options = {}) {
+  const hasOnSanitize = typeof options.onSanitize === 'function';
   return function (req, res, next) {
-    ['body', 'params', 'headers', 'query'].forEach(function (k) {
-      if (req[k]) {
-        req[k] = sanitize(req[k], options);
+    ['body', 'params', 'headers', 'query'].forEach(function (key) {
+      if (req[key]) {
+        const { target, isSanitized } = _sanitize(req[key], options);
+        req[key] = target;
+        if (isSanitized && hasOnSanitize) {
+          options.onSanitize({
+            req,
+            key,
+          });
+        }
       }
     });
     next();