+++ test.js
@@ -5,6 +5,17 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const expect = require('chai').expect;
 const sanitize = require('./index.js');
+const callTracker = () => {
+  const callStack = [];
+  return {
+    report() {
+      return callStack;
+    },
+    call(...args) {
+      callStack.push(args);
+    },
+  };
+};
 
 describe('Express Mongo Sanitize', function () {
   describe('Remove Data', function () {
@@ -43,7 +54,7 @@ describe('Express Mongo Sanitize', function () {
                 q: 'search',
               },
             },
-            done
+            done,
           );
       });
 
@@ -71,7 +82,7 @@ describe('Express Mongo Sanitize', function () {
                 even: null,
               },
             },
-            done
+            done,
           );
       });
 
@@ -111,7 +122,7 @@ describe('Express Mongo Sanitize', function () {
                 q: 'search',
               },
             },
-            done
+            done,
           );
       });
     });
@@ -128,7 +139,7 @@ describe('Express Mongo Sanitize', function () {
                 username: {},
               },
             },
-            done
+            done,
           );
       });
 
@@ -150,7 +161,7 @@ describe('Express Mongo Sanitize', function () {
                 username: {},
               },
             },
-            done
+            done,
           );
       });
 
@@ -167,7 +178,7 @@ describe('Express Mongo Sanitize', function () {
                 username: {},
               },
             },
-            done
+            done,
           );
       });
     });
@@ -184,7 +195,7 @@ describe('Express Mongo Sanitize', function () {
                 username: [{}],
               },
             },
-            done
+            done,
           );
       });
 
@@ -208,7 +219,7 @@ describe('Express Mongo Sanitize', function () {
                 username: [{}],
               },
             },
-            done
+            done,
           );
       });
 
@@ -225,7 +236,7 @@ describe('Express Mongo Sanitize', function () {
                 username: [{}],
               },
             },
-            done
+            done,
           );
       });
     });
@@ -238,7 +249,7 @@ describe('Express Mongo Sanitize', function () {
     app.use(
       sanitize({
         replaceWith: '_',
-      })
+      }),
     );
 
     app.post('/body', function (req, res) {
@@ -273,7 +284,7 @@ describe('Express Mongo Sanitize', function () {
                 dotted_data: 'some_data',
               },
             },
-            done
+            done,
           );
       });
 
@@ -303,7 +314,7 @@ describe('Express Mongo Sanitize', function () {
                 dotted_data: 'some_data',
               },
             },
-            done
+            done,
           );
       });
 
@@ -346,7 +357,7 @@ describe('Express Mongo Sanitize', function () {
                 dotted_data: 'some_data',
               },
             },
-            done
+            done,
           );
       });
     });
@@ -366,7 +377,7 @@ describe('Express Mongo Sanitize', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
 
@@ -391,7 +402,7 @@ describe('Express Mongo Sanitize', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
 
@@ -411,7 +422,7 @@ describe('Express Mongo Sanitize', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
     });
@@ -433,7 +444,7 @@ describe('Express Mongo Sanitize', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
 
@@ -462,7 +473,7 @@ describe('Express Mongo Sanitize', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
 
@@ -484,7 +495,7 @@ describe('Express Mongo Sanitize', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
     });
@@ -515,7 +526,7 @@ describe('Express Mongo Sanitize', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
     });
@@ -562,7 +573,7 @@ describe('Express Mongo Sanitize', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
       it('should not set constructor property', function (done) {
@@ -591,7 +602,7 @@ describe('Express Mongo Sanitize', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
       it('should not set prototype property', function (done) {
@@ -620,7 +631,7 @@ describe('Express Mongo Sanitize', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
     });
@@ -633,7 +644,7 @@ describe('Express Mongo Sanitize', function () {
       app.use(
         sanitize({
           replaceWith: '$',
-        })
+        }),
       );
 
       app.get('/query', function (req, res) {
@@ -651,7 +662,7 @@ describe('Express Mongo Sanitize', function () {
               q: 'search',
             },
           },
-          done
+          done,
         );
     });
 
@@ -661,7 +672,7 @@ describe('Express Mongo Sanitize', function () {
       app.use(
         sanitize({
           replaceWith: '.',
-        })
+        }),
       );
 
       app.get('/query', function (req, res) {
@@ -679,7 +690,7 @@ describe('Express Mongo Sanitize', function () {
               q: 'search',
             },
           },
-          done
+          done,
         );
     });
   });
@@ -782,6 +793,191 @@ describe('Express Mongo Sanitize', function () {
       expect(sanitize.has(input)).to.be.false;
     });
   });
+
+  describe('onSanitize', function () {
+    it('should not call onSanitize if the object is valid', function (done) {
+      const tracker = callTracker();
+      const app = express();
+      app.use(bodyParser.urlencoded({ extended: true }));
+      app.use(bodyParser.json());
+      app.use(
+        sanitize({
+          dryRun: true,
+          onSanitize(arg) {
+            tracker.call(arg);
+          },
+        }),
+      );
+      app.post('/body', function (req, res) {
+        res.status(200).json({
+          body: req.body,
+        });
+      });
+      request(app)
+        .post('/body')
+        .send({
+          q: 'search',
+        })
+        .set('Content-Type', 'application/json')
+        .set('Accept', 'application/json')
+        .expect(
+          200,
+          {
+            body: {
+              q: 'search',
+            },
+          },
+          (error) => {
+            expect(tracker.report()).to.have.lengthOf(0);
+            done(error);
+          },
+        );
+    });
+    it('should call onSanitize if the object has a malicious key', function (done) {
+      const tracker = callTracker();
+      const app = express();
+      app.use(bodyParser.urlencoded({ extended: true }));
+      app.use(bodyParser.json());
+      app.use(
+        sanitize({
+          onSanitize(arg) {
+            tracker.call(arg);
+          },
+        }),
+      );
+      app.post('/body', function (req, res) {
+        res.status(200).json({
+          body: req.body,
+        });
+      });
+      request(app)
+        .post('/body')
+        .send({
+          q: 'search',
+          is: true,
+          and: 1,
+          even: null,
+          stop: undefined,
+          $where: 'malicious',
+          'dotted.data': 'some_data',
+        })
+        .set('Content-Type', 'application/json')
+        .set('Accept', 'application/json')
+        .expect(
+          200,
+          {
+            body: {
+              q: 'search',
+              is: true,
+              and: 1,
+              even: null,
+            },
+          },
+          (error) => {
+            const callStack = tracker.report();
+            expect(callStack).to.have.lengthOf(1);
+            const calledFirstArgument = callStack[0][0];
+            expect(calledFirstArgument).to.have.property('req');
+            expect(calledFirstArgument).to.have.property('key', 'body');
+            done(error);
+          },
+        );
+    });
+  });
+  describe('dryRun', function () {
+    it('should not sanitized if the object has a malicious key', function (done) {
+      const app = express();
+      app.use(bodyParser.urlencoded({ extended: true }));
+      app.use(bodyParser.json());
+      app.use(
+        sanitize({
+          dryRun: true,
+        }),
+      );
+      app.post('/body', function (req, res) {
+        res.status(200).json({
+          body: req.body,
+        });
+      });
+      request(app)
+        .post('/body')
+        .send({
+          q: 'search',
+          is: true,
+          and: 1,
+          even: null,
+          $where: 'malicious',
+          'dotted.data': 'some_data',
+        })
+        .set('Content-Type', 'application/json')
+        .set('Accept', 'application/json')
+        .expect(
+          200,
+          {
+            body: {
+              q: 'search',
+              is: true,
+              and: 1,
+              even: null,
+              $where: 'malicious',
+              'dotted.data': 'some_data',
+            },
+          },
+          done,
+        );
+    });
+    it('should not sanitized, but call onSanitize if the object has a malicious key', function (done) {
+      const tracker = callTracker();
+      const app = express();
+      app.use(bodyParser.urlencoded({ extended: true }));
+      app.use(bodyParser.json());
+      app.use(
+        sanitize({
+          dryRun: true,
+          onSanitize(arg) {
+            tracker.call(arg);
+          },
+        }),
+      );
+      app.post('/body', function (req, res) {
+        res.status(200).json({
+          body: req.body,
+        });
+      });
+      request(app)
+        .post('/body')
+        .send({
+          q: 'search',
+          is: true,
+          and: 1,
+          even: null,
+          $where: 'malicious',
+          'dotted.data': 'some_data',
+        })
+        .set('Content-Type', 'application/json')
+        .set('Accept', 'application/json')
+        .expect(
+          200,
+          {
+            body: {
+              q: 'search',
+              is: true,
+              and: 1,
+              even: null,
+              $where: 'malicious',
+              'dotted.data': 'some_data',
+            },
+          },
+          (error) => {
+            const callStack = tracker.report();
+            expect(callStack).to.have.lengthOf(1);
+            const calledFirstArgument = callStack[0][0];
+            expect(calledFirstArgument).to.have.property('key', 'body');
+            done(error);
+          },
+        );
+    });
+  });
 });
 
 describe('Express Mongo Sanitize, Dots included', function () {
@@ -822,7 +1018,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 'dotted.data': 'some_data',
               },
             },
-            done
+            done,
           );
       });
 
@@ -851,7 +1047,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 'dotted.data': 'some_data',
               },
             },
-            done
+            done,
           );
       });
 
@@ -893,7 +1089,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 'dotted.data': 'some_data',
               },
             },
-            done
+            done,
           );
       });
     });
@@ -912,7 +1108,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
 
@@ -936,7 +1132,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
 
@@ -955,7 +1151,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
     });
@@ -976,7 +1172,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
 
@@ -1004,7 +1200,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
 
@@ -1025,7 +1221,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
     });
@@ -1039,7 +1235,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
       sanitize({
         replaceWith: '_',
         allowDots: true,
-      })
+      }),
     );
 
     app.post('/body', function (req, res) {
@@ -1074,7 +1270,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 'dotted.data': 'some_data',
               },
             },
-            done
+            done,
           );
       });
 
@@ -1104,7 +1300,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 'dotted.data': 'some_data',
               },
             },
-            done
+            done,
           );
       });
 
@@ -1147,7 +1343,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 'dotted.data': 'some_data',
               },
             },
-            done
+            done,
           );
       });
     });
@@ -1167,7 +1363,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
 
@@ -1192,7 +1388,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
 
@@ -1212,7 +1408,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
     });
@@ -1234,7 +1430,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
 
@@ -1263,7 +1459,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
 
@@ -1285,7 +1481,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 ],
               },
             },
-            done
+            done,
           );
       });
     });
@@ -1316,7 +1512,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
     });
@@ -1364,7 +1560,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
       it('should not set constructor property', function (done) {
@@ -1394,7 +1590,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
       it('should not set prototype property', function (done) {
@@ -1424,7 +1620,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
                 },
               },
             },
-            done
+            done,
           );
       });
     });
@@ -1438,7 +1634,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
         sanitize({
           replaceWith: '$',
           allowDots: true,
-        })
+        }),
       );
 
       app.get('/query', function (req, res) {
@@ -1457,7 +1653,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
               'dotted.data': 'some_data',
             },
           },
-          done
+          done,
         );
     });
 
@@ -1468,7 +1664,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
         sanitize({
           replaceWith: '.',
           allowDots: true,
-        })
+        }),
       );
 
       app.get('/query', function (req, res) {
@@ -1487,7 +1683,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
               'dotted.data': 'some_data',
             },
           },
-          done
+          done,
         );
     });
   });
@@ -1500,7 +1696,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
       expect(sanitize.has(input, true)).to.be.true;
     });
 
-    it('should return false if the object has a key containing a `.`, when allowDots=true', function () {
+    it('should return false if the object has a key containing a `.`, but only when allowDots=true', function () {
       const input = {
         'prohibited.key': 'value',
       };
@@ -1516,7 +1712,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
       expect(sanitize.has(input, true)).to.be.true;
     });
 
-    it('should return true if the object has a nested key containing a `.`, when allowDots=true', function () {
+    it('should return true if the object has a nested key containing a `.`, but only when allowDots=true', function () {
       const input = {
         nested: {
           'prohibited.key': 'value',
@@ -1534,7 +1730,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
       expect(sanitize.has(input, true)).to.be.true;
     });
 
-    it('should return true if the array contains an object with a key containing a `.`, when allowDots=true', function () {
+    it('should return true if the array contains an object with a key containing a `.`, but only when allowDots=true', function () {
       const input = [
         {
           'prohibited.key': 'value',
@@ -1560,7 +1756,7 @@ describe('Express Mongo Sanitize, Dots included', function () {
       expect(sanitize.has(input, true)).to.be.true;
     });
 
-    it('should return true if the payload contains a deeply nested object with a key containing a `.`, when allowDots=true', function () {
+    it('should return true if the payload contains a deeply nested object with a key containing a `.`, but only when allowDots=true', function () {
       const input = [
         {
           some: {